requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.1
python-dotenv==1.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
fake-useragent==1.4.0
lxml==4.9.3
gunicorn==21.2.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
pywebpush==1.14.0
supabase==2.4.0
psycopg2-binary==2.9.9 